+++ app/src/main/java/com/growthook/aos/presentation/cavedetail/CaveDetailCaveSelectBottomSheet.kt
@@ -14,7 +14,7 @@ import androidx.recyclerview.selection.StableIdKeyProvider
 import androidx.recyclerview.selection.StorageStrategy
 import com.growthook.aos.R
 import com.growthook.aos.databinding.FragmentCaveSelectBottomsheetBinding
-import com.growthook.aos.presentation.home.CaveSelectAdapter
+import com.growthook.aos.util.selectcave.CaveSelectAdapter
 import com.growthook.aos.util.base.BaseBottomSheetFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
